python/requirement.txt
selenium==3.141.0
pytest==5.2.0
pytest-xdist==1.24.1
Faker==1.0.5
webdriver_manager==1.7
allure-pytest==2.8.13
Appium-Python-Client==0.43
pytest-testrail==2.8.3
rootpath==0.1.1